feat: Added Databases abstractions.
