+++ src/Databases/Abstractions/src/ESearchType.cs
