+++ src/Databases/Chroma/src/LangChain.Databases.Chroma.csproj
@@ -19,7 +19,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\..\Core\src\LangChain.Core.csproj" />
+      <ProjectReference Include="..\..\Abstractions\src\LangChain.Databases.Abstractions.csproj" />
     </ItemGroup>
 
 </Project>
