+++ src/Databases/Postgres/src/LangChain.Databases.Postgres.csproj
@@ -17,7 +17,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\..\Core\src\LangChain.Core.csproj" />
+        <ProjectReference Include="..\..\Abstractions\src\LangChain.Databases.Abstractions.csproj" />
     </ItemGroup>
 
 </Project>
