+++ src/Databases/Sqlite/src/LangChain.Databases.Sqlite.csproj
@@ -16,7 +16,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\..\Core\src\LangChain.Core.csproj" />
+        <ProjectReference Include="..\..\Abstractions\src\LangChain.Databases.Abstractions.csproj" />
     </ItemGroup>
 
 </Project>
